fix: extract event-feed test fixtures to satisfy 800-line gate (#4146)
